+++ src/scad_tarmi/threaded.clj
@@ -5,95 +5,8 @@
             [scad-clj.model :as model]
             [scad-tarmi.core :refer [sin cos τ long-hex-diagonal]]
             [scad-tarmi.maybe :as maybe]
-            [scad-tarmi.dfm :as dfm]))
-
-;;;;;;;;;;;;;;;;;;;;;;;;
-;; INTERNAL CONSTANTS ;;
-;;;;;;;;;;;;;;;;;;;;;;;;
-
-(spec/def ::iso-property #{:hex-head-short-diagonal
-                           :hex-head-long-diagonal
-                           :head-hex-drive-short-diagonal
-                           :head-hex-drive-long-diagonal
-                           :iso4017-hex-head-height-nominal
-                           :hex-nut-height
-                           :socket-diameter
-                           :socket-height
-                           :button-diameter
-                           :button-height
-                           :countersunk-diameter
-                           :countersunk-height
-                           :thread-pitch-coarse
-                           :iso7089-inner-diameter
-                           :iso7089-outer-diameter
-                           :iso7089-thickness})
-
-(def ^:internal iso-data
-  "Various constants from ISO metric fastener standards.
-  This is a map of nominal ISO bolt diameter (in mm) to various other
-  measurements according to spec. Instead of relying on this raw data in
-  applications, prefer the more capable datum function."
-  {3 {:socket-diameter 5.5
-      :hex-nut-height 2.4
-      :iso4017-hex-head-height-nominal 2
-      :thread-pitch-coarse 0.5
-      :head-hex-drive-short-diagonal 2.5
-      :iso7089-inner-diameter 3.2
-      :iso7089-outer-diameter 7
-      :iso7089-thickness 0.5}
-   4 {:socket-diameter 7
-      :hex-nut-height 3.2
-      :iso4017-hex-head-height-nominal 2.8
-      :thread-pitch-coarse 0.7
-      :head-hex-drive-short-diagonal 3
-      :iso7089-inner-diameter 4.3
-      :iso7089-outer-diameter 9
-      :iso7089-thickness 0.8}
-   5 {:socket-diameter 8.5
-      :hex-head-short-diagonal 8
-      :hex-nut-height 4.7
-      :iso4017-hex-head-height-nominal 3.5
-      :thread-pitch-coarse 0.8
-      :head-hex-drive-short-diagonal 4
-      :iso7089-inner-diameter 5.3
-      :iso7089-outer-diameter 10
-      :iso7089-thickness 1}
-   6 {:socket-diameter 10
-      :hex-nut-height 5.2
-      :iso4017-hex-head-height-nominal 4
-      :thread-pitch-coarse 1
-      :head-hex-drive-short-diagonal 5
-      :iso7089-inner-diameter 6.4
-      :iso7089-outer-diameter 12
-      :iso7089-thickness 1.6}
-   8 {:socket-diameter 13
-      :hex-nut-height 6.8
-      :iso4017-hex-head-height-nominal 5.3
-      :thread-pitch-coarse 1.25
-      :head-hex-drive-short-diagonal 6
-      :iso7089-inner-diameter 8.4
-      :iso7089-outer-diameter 16
-      :iso7089-thickness 1.6}})
-
-
-;;;;;;;;;;;;;;;;;;;;;
-;; INTERFACE SPECS ;;
-;;;;;;;;;;;;;;;;;;;;;
-
-;; The following items are exposed for use in application data validation.
-
-;; Supported nominal diameters:
-(spec/def ::iso-nominal #(contains? iso-data %))
-
-;; Supported types of bolt heads:
-(spec/def ::head-type #{:hex     ; Hex head with the diameter of a nut.
-                        :socket  ; Full cylindrical counterbore cap.
-                        :button  ; Partial (low, smooth-edged) socket cap.
-                        :countersunk}) ; Flat head tapering toward the bolt.
-
-;; Supported types of bolt drives and points:
-(spec/def ::drive-type #{:hex})
-(spec/def ::point-type #{:cone})
+            [scad-tarmi.dfm :as dfm]
+            [scad-tarmi.threaded.schema :as schema]))
 
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
@@ -103,9 +16,9 @@
 (defn datum
   "Retrieve or calculate a fact based on the ISO standards."
   [nominal-diameter key]
-  {:pre [(spec/valid? ::iso-nominal nominal-diameter)
-         (spec/valid? ::iso-property key)]}
-  (let [data (get iso-data nominal-diameter)]
+  {:pre [(spec/valid? ::schema/iso-size nominal-diameter)
+         (spec/valid? ::schema/iso-property key)]}
+  (let [data (get schema/iso-data nominal-diameter)]
    (case key
      :hex-head-short-diagonal  ; Flat-to-flat width of a hex head.
        ;; For most sizes, this value is equal to socket diameter.
@@ -139,8 +52,8 @@
   This is exposed for predicting the results of the bolt function in this
   module, specifically where the transition from head to body will occur."
   [iso-size head-type]
-  {:pre [(spec/valid? ::iso-nominal iso-size)
-         (spec/valid? ::head-type head-type)]}
+  {:pre [(spec/valid? ::schema/iso-size iso-size)
+         (spec/valid? ::schema/head-type head-type)]}
   (datum iso-size
     (case head-type
       :hex :iso4017-hex-head-height-nominal
@@ -197,8 +110,8 @@
   to make sure the head will not protrude with normal printing defects."
   [{:keys [iso-size head-type countersink-edge-fn compensator]
     :or {countersink-edge-fn (fn [iso-size] (/ (Math/log iso-size) 8))}}]
-  {:pre [(spec/valid? ::iso-nominal iso-size)
-         (spec/valid? ::head-type head-type)]}
+  {:pre [(spec/valid? ::schema/iso-size iso-size)
+         (spec/valid? ::schema/head-type head-type)]}
   (let [height (head-height iso-size head-type)]
     (case head-type
       :hex
@@ -222,8 +135,8 @@
 (defn- bolt-drive
   "A model of the thing you stick your bit in."
   [{:keys [iso-size head-type drive-type drive-recess-depth]}]
-  {:pre [(spec/valid? ::iso-nominal iso-size)
-         (spec/valid? ::drive-type drive-type)]}
+  {:pre [(spec/valid? ::schema/iso-size iso-size)
+         (spec/valid? ::schema/drive-type drive-type)]}
   (let [depth (or drive-recess-depth
                   (/ (head-height iso-size head-type) 2))]
     (model/translate [0 0 (/ depth -2)]
@@ -322,7 +235,7 @@
 
 (defn- cone
   "A trivial conical point to a bolt. The characteristics of the cone
-  follow ISO 7434 (i.e. 90º angle, assuming a long screw; the tip is not
+  follow ISO 7434 (i.e. 90° angle, assuming a long screw; the tip is not
   blunted), but it’s not intended for making slotted set screws, just to
   prevent slicing software from building support inside negatives in the
   bottom of an object, where support is difficult to remove."
@@ -406,7 +319,7 @@
       :or {taper-fn rounding-taper, threaded true, compensator dfm/none,
            negative false}
       :as options}]
-  {:pre [(spec/valid? ::iso-nominal iso-size)]}
+  {:pre [(spec/valid? ::schema/iso-size iso-size)]}
   (let [options (merge {:outer-diameter iso-size
                         :pitch (datum iso-size :thread-pitch-coarse)
                         :taper-fn taper-fn}
@@ -428,14 +341,11 @@
   drive is supported, and even that will be ignored on a negative.
   Likewise, though a point-type parameter is accepted, the only implemented
   option beyond the default flat point is a cone."
-  [& {:keys [iso-size pitch head-type drive-type point-type threaded
-             total-length unthreaded-length threaded-length
-             compensator negative]
-      :or {head-type :hex, threaded true, compensator dfm/none, negative false}
+  [& {:keys [iso-size pitch total-length unthreaded-length threaded-length
+             head-type drive-type point-type threaded negative compensator]
+      :or {head-type :hex, threaded true, negative false, compensator dfm/none}
       :as options}]
-  {:pre [(spec/valid? ::iso-nominal iso-size)
-         (spec/valid? (spec/nilable ::head-type) head-type)
-         (spec/valid? (spec/nilable ::point-type) point-type)]}
+  {:pre [(spec/valid? ::schema/bolt-parameters options)]}
   (let [hh (head-height iso-size head-type)
         pitch (or pitch (datum iso-size :thread-pitch-coarse))
         lengths (bolt-length
@@ -485,7 +395,7 @@
   "A single hex nut centred at [0 0 0]."
   [& {:keys [iso-size height compensator negative]
       :or {compensator dfm/none}}]
-  {:pre [(spec/valid? ::iso-nominal iso-size)]}
+  {:pre [(spec/valid? ::schema/iso-size iso-size)]}
   (let [height (or height (datum iso-size :hex-nut-height))]
     (if negative
       ;; A convex model of a nut.
@@ -510,3 +420,15 @@
     (model/difference
       (model/cylinder (/ od 2) thickness)
       (model/cylinder (/ id 2) (+ thickness 1)))))
+
+
+;;;;;;;;;;;;;;;;
+;; DEPRECATED ;;
+;;;;;;;;;;;;;;;;
+
+;; The following items are deprecated since scad-tarmi v0.5.0 and will be removed in v1.0.0.
+
+(spec/def ::iso-nominal ::schema/iso-size)
+(spec/def ::head-type ::schema/head-type)
+(spec/def ::drive-type ::schema/drive-type)
+(spec/def ::point-type ::schema/point-type)